Add keyField and refField to database generics
